Populate upnp devices from ssdp
