+++ homeassistant/components/upnp/__init__.py
@@ -5,6 +5,7 @@ from ipaddress import ip_address
 import voluptuous as vol
 
 from homeassistant import config_entries
+from homeassistant.components import ssdp
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
@@ -17,9 +18,6 @@ from .const import (
     CONFIG_ENTRY_HOSTNAME,
     CONFIG_ENTRY_ST,
     CONFIG_ENTRY_UDN,
-    DISCOVERY_LOCATION,
-    DISCOVERY_ST,
-    DISCOVERY_UDN,
     DOMAIN,
     DOMAIN_CONFIG,
     DOMAIN_DEVICES,
@@ -49,24 +47,15 @@ async def async_construct_device(hass: HomeAssistant, udn: str, st: str) -> Devi
     """Discovery devices and construct a Device for one."""
     # pylint: disable=invalid-name
     _LOGGER.debug("Constructing device: %s::%s", udn, st)
+    discovery_info = ssdp.async_get_discovery_info_by_udn_st(hass, udn, st)
 
-    discoveries = [
-        discovery
-        for discovery in await Device.async_discover(hass)
-        if discovery[DISCOVERY_UDN] == udn and discovery[DISCOVERY_ST] == st
-    ]
-    if not discoveries:
+    if not discovery_info:
         _LOGGER.info("Device not discovered")
         return None
 
-    # Some additional clues for remote debugging.
-    if len(discoveries) > 1:
-        _LOGGER.info("Multiple devices discovered: %s", discoveries)
-
-    discovery = discoveries[0]
-    _LOGGER.debug("Constructing from discovery: %s", discovery)
-    location = discovery[DISCOVERY_LOCATION]
-    return await Device.async_create_device(hass, location)
+    return await Device.async_create_device(
+        hass, discovery_info[ssdp.ATTR_SSDP_LOCATION]
+    )
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType):
