+++ homeassistant/components/upnp/config_flow.py
@@ -29,17 +29,7 @@ from .const import (
     DOMAIN_DEVICES,
     LOGGER as _LOGGER,
 )
-from .device import Device
-
-
-def discovery_info_to_discovery(discovery_info: Mapping) -> Mapping:
-    """Convert a SSDP-discovery to 'our' discovery."""
-    return {
-        DISCOVERY_UDN: discovery_info[ssdp.ATTR_UPNP_UDN],
-        DISCOVERY_ST: discovery_info[ssdp.ATTR_SSDP_ST],
-        DISCOVERY_LOCATION: discovery_info[ssdp.ATTR_SSDP_LOCATION],
-        DISCOVERY_USN: discovery_info[ssdp.ATTR_SSDP_USN],
-    }
+from .device import Device, discovery_info_to_discovery
 
 
 class UpnpFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
