+++ homeassistant/components/upnp/manifest.json
@@ -4,6 +4,7 @@
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/upnp",
   "requirements": ["async-upnp-client==0.18.0"],
+  "dependencies": ["ssdp"],
   "codeowners": ["@StevenLooman"],
   "ssdp": [
     {
